Refactor to extensions for span position calculations
