+++ Xamarin.Forms.Platform.UAP/Xamarin.Forms.Platform.UAP.csproj
@@ -64,6 +64,7 @@
     <Compile Include="NativeViewWrapperRenderer.cs" />
     <Compile Include="PlatformConfigurationExtensions.cs" />
     <Compile Include="PlatformEffect.cs" />
+    <Compile Include="TextBlockExtensions.cs" />
     <Compile Include="UriImageSourceHandler.cs" />
     <Compile Include="ViewExtensions.cs" />
     <Compile Include="LayoutExtensions.cs" />
