+++ Xamarin.Forms.Platform.WPF/Xamarin.Forms.Platform.WPF.csproj
@@ -19,7 +19,7 @@
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
     <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <DefineConstants>TRACE;DEBUG</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <LangVersion>6</LangVersion>
@@ -62,6 +62,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\Xamarin.Forms.Platform.UAP\TextBlockExtensions.cs">
+      <Link>Extensions\TextBlockExtensions.cs</Link>
+    </Compile>
     <Compile Include="Animatable.cs" />
     <Compile Include="CellControl.cs" />
     <Compile Include="Converters\CaseConverter.cs" />
