+++ Xamarin.Forms.Platform.iOS/Renderers/LabelRenderer.cs
@@ -110,128 +110,9 @@ namespace Xamarin.Forms.Platform.MacOS
 					break;
 			}
 
-			RecalculateSpanPositions(Control.Frame);
+			Control.RecalculateSpanPositions(Element);
 
-		}
-
-		double FindDefaultLineHeight(int start, int length)
-		{
-			if (length == 0)
-				return 0.0;
-
-			var textStorage = new NSTextStorage();
-#if __MOBILE__
-			textStorage.SetString(Control.AttributedText.Substring(start, length));
-#else
-			textStorage.SetString(Control.AttributedStringValue.Substring(start, length));
-#endif
-			var layoutManager = new NSLayoutManager();
-			textStorage.AddLayoutManager(layoutManager);
-
-			var textContainer = new NSTextContainer(size: new SizeF(double.MaxValue, double.MaxValue))
-			{
-				LineFragmentPadding = 0
-			};
-			layoutManager.AddTextContainer(textContainer);
-
-			var glyph = new NSRange();
-#if __MOBILE__
-			layoutManager.CharacterRangeForGlyphRange(new NSRange(0, 1), ref glyph);
-#else
-			layoutManager.CharacterRangeForGlyphRange(new NSRange(0, 1), out glyph);
-#endif
-			var rect = layoutManager.BoundingRectForGlyphRange(glyph, textContainer);
-			return rect.Bottom - rect.Top;
-		}
-
-		void RecalculateSpanPositions(RectangleF finalSize)
-		{
-			if (Element?.FormattedText?.Spans == null
-				|| Element.FormattedText.Spans.Count == 0)
-				return;
-
-			if (finalSize.Width <= 0 || finalSize.Height <= 0)
-				return;
-
-#if __MOBILE__
-			var inline = Control.AttributedText;
-#else
-			var inline = Control.AttributedStringValue;
-#endif
-			var range = new NSRange(0, inline.Length);
-
-			NSTextStorage textStorage = new NSTextStorage();
-			textStorage.SetString(inline);
-
-			var layoutManager = new NSLayoutManager();
-
-			textStorage.AddLayoutManager(layoutManager);
-
-			var textContainer = new NSTextContainer(size: Control.Frame.Size)
-			{
-				LineFragmentPadding = 0
-			};
-
-			layoutManager.AddTextContainer(textContainer);
-
-			var labelWidth = finalSize.Width;
-
-			var currentLocation = 0;
-
-			for (int i = 0; i < Element.FormattedText.Spans.Count; i++)
-			{
-				var span = Element.FormattedText.Spans[i];
-
-				var location = currentLocation;
-				var length = span.Text.Length;
-
-				var startRect = GetCharacterBounds(new NSRange(location, 1), layoutManager, textContainer);
-				var endRect = GetCharacterBounds(new NSRange(location + length, 1), layoutManager, textContainer);
-
-				var startLineHeight = startRect.Bottom - startRect.Top;
-				var endLineHeight = endRect.Bottom - endRect.Top;
-
-				var defaultLineHeight = FindDefaultLineHeight(location, length);
-
-				var yaxis = startRect.Top;
-				var lineHeights = new List<double>();
-				while (yaxis < endRect.Bottom)
-				{
-					double lineHeight;
-					if (yaxis == startRect.Top) // First Line
-					{
-						lineHeight = startRect.Bottom - startRect.Top;
-					}
-					else if (yaxis != endRect.Top) // Middle Line(s)
-					{
-						lineHeight = defaultLineHeight;
-					}
-					else // Bottom Line
-					{
-						lineHeight = endRect.Bottom - endRect.Top;
-					}
-					lineHeights.Add(lineHeight);
-					yaxis += (nfloat)lineHeight;
-				}
-
-				span.CalculatePositions(lineHeights.ToArray(), finalSize.Width, startRect.X, endRect.X, startRect.Top);
-
-				// update current location
-				currentLocation += length;
-			}
-		}
-
-		RectangleF GetCharacterBounds(NSRange characterRange, NSLayoutManager layoutManager, NSTextContainer textContainer)
-		{
-			var glyphRange = new NSRange();
-
-#if __MOBILE__
-			layoutManager.CharacterRangeForGlyphRange(characterRange, ref glyphRange);
-#else
-			layoutManager.CharacterRangeForGlyphRange(characterRange, out glyphRange);
-#endif
-			return layoutManager.BoundingRectForGlyphRange(glyphRange, textContainer);
-		}
+		}		
 
 		protected override void OnElementChanged(ElementChangedEventArgs<Label> e)
 		{
